requirements.txt
apscheduler==3.9.1
beautifulsoup4==4.11.1
openpyxl==3.0.10
python_dateutil==2.8.2
requests==2.27.1
tqdm==4.64.1
urllib3==1.26.12
